+++ src/unstoppableAPI.ts
@@ -1,7 +1,9 @@
+import { toBech32Address } from '@zilliqa-js/crypto';
+import nodeFetch from 'node-fetch';
+
 import { ResolutionError } from './index';
 import NamingService from './namingService';
-import nodeFetch from 'node-fetch';
-import { NamicornResolution, NullAddress, SourceDefinition, NamingServiceSource } from './types';
+import { NamicornResolution, NullAddress, NamingServiceSource, SourceDefinition } from './types';
 import Zns from './zns';
 import Ens from './ens';
 import fs from 'fs';
@@ -36,7 +38,7 @@ export default class Udapi extends NamingService {
     const DefaultUserAgent = isNode()
       ? 'node-fetch/1.0 (+https://github.com/bitinn/node-fetch)'
       : navigator.userAgent;
-    const pckg = JSON.parse(fs.readFileSync('./package.json') as any);
+    const pckg = JSON.parse(fs.readFileSync(`${__dirname}/../package.json`) as any);
     const version = pckg.version;
     const CustomUserAgent = `${DefaultUserAgent} namicorn/${version}`;
     this.headers = { 'X-user-agent': CustomUserAgent };
@@ -77,6 +79,17 @@ export default class Udapi extends NamingService {
         currencyTicker,
       });
     return address;
+  }
+
+  /**
+   * Owner of the domain
+   * @param domain - domain name
+   * @returns - an owner address of the domain
+   */
+  async owner(domain: string): Promise<string | null> {
+    const {owner} = (await this.resolve(domain)).meta;
+    if (!owner) return null;
+    return owner.startsWith("zil1") ? owner : toBech32Address(owner);
   }
 
   /**