+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+* Namicorn#owner method - returns an owner address of the domain
+
 ## 0.3.0
 * Namicorn#addressOrThrow - new method that throws ResolutionError if currency address is not found
 * Namicorn#namehash - new method for namehashing the domain name. Name hash of a domain is an ID that is used to store the information about the domain on the blockchain. If you would browse the blockchain, you would never see domain names, just name hashes.